Add checks for Dotenv before calling it
`composer install --no-dev` fails when Dotenv is a dev dependency (which
it is in skeleton) and you're running in an environment without an
APP_ENV variable set.

Throw exception when no env variables are set and Dotenv is absent so
the user knows to either require DotEnv or specify an APP_ENV variable.